--- tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/weather/special_weather/SpecialWeatherDetail.java
@@ -1,13 +1,0 @@
-package kr.ac.kumoh.illdang100.tovalley.domain.weather.special_weather;
-
-import lombok.*;
-
-
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SpecialWeatherDetail {
-
-    private String content;
-}
